libhb: nlmeans: add high depth support.
